refactor: extract indent validation functions into dedicated module
